Rename directory to match API
