+++ web-common/src/features/metrics-views/DimensionColumns.ts
