+++ web-common/src/features/metrics-views/metrics-internal-store.ts
