+++ web-common/src/features/metrics-views/workspace/MetricsAddEntityButton.svelte
