+++ web-common/src/features/metrics-views/workspace/MetricsEntityTable.svelte
