+++ web-common/src/features/metrics-views/workspace/MetricsGenerateButton.svelte
