+++ web-common/src/features/metrics-views/workspace/MetricsLayoutManager.svelte
