+++ web-common/src/features/metrics-views/workspace/QuickMetricsModal.svelte
